boards: tt_blackhole: simplify bmc p150 overlay file
The bmc_p150 overlay file contained a number of redundant nodes
(i.e. nodes that did not differ from those in the .dts files
that they overlaid on top of).

* model and compatible in `/` node identical
* `spi1` phandle not needed to get `flash1` phandle
* `flash0` partitions identical
* `i2c1` node identical
* `smbus1` node identical

Other than that, the only differences that need to be preserved
in the bmc p150 overlay file are `jedec-id` field of `flash1`.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>