boards: tenstorrent: tt_blackhole: add p100a board revision
Add a P100A board revision which will be the board that ships to
customers.

The P100 board revision is mainly used for internal purposes.

The P100 and P100A board revisions are nearly identical from the
perspective of firmware. The only difference being that on the
P100, pci0 is deactivated (at runtime) via the fw_table, while
on the P100A, pci1 is deactivated (at runtime) via the fw_table.

So P100 -> pci1, P100A -> pci0.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>