tenstorrent: bh_chip: add pgood fault interrupt
Adds interrupts on pgood gpio fall and rise to enable the pgood fault sequence.

When a pgood fault is detected, the board fault LED is turned on and ASIC reset
is asserted. If pgood comes back up, then the LED is cleared and the ASIC is
brought out of reset. If it fails to rise again, the pgood fault state can
only be cleared by a power cycle.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>